+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [7.4.1 (unreleased)](https://github.com/kdeldycke/click-extra/compare/v7.4.0...main)
+## [7.5.0 (unreleased)](https://github.com/kdeldycke/click-extra/compare/v7.4.0...main)
 
 > [!IMPORTANT]
 > This version is not released yet and is under active development.